added derive_operatorid command to egnkey also refactored commands to be in their own packages to prevent var/flag conflicts
