+++ cmd/egnkey/main.go
@@ -4,6 +4,9 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/Layr-Labs/eigensdk-go/cmd/egnkey/derive_operatorid"
+	"github.com/Layr-Labs/eigensdk-go/cmd/egnkey/generate"
+	"github.com/Layr-Labs/eigensdk-go/cmd/egnkey/store"
 	"github.com/urfave/cli/v2"
 )
 
@@ -12,8 +15,9 @@ func main() {
 	app.Name = "egnkey"
 	app.Description = "Eigenlayer batch keys manager"
 	app.Commands = []*cli.Command{
-		commandGenerate,
-		commandStore,
+		generate.Command,
+		store.Command,
+		derive_operatorid.Command,
 	}
 
 	app.Usage = "Used to manage batch keys for testing"
